database v.1

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -20,7 +19,7 @@ return new class extends Migration
             $table->string('password');
             $table->string('alamat')->nullable();
             $table->enum('role', ['Admin', 'Petugas', 'Warga'])->nullable();
-            $table->integer('nomor')->nullable();
+            $table->bigInteger('nomor')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->rememberToken();
             $table->timestamps();
